+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightLoop/LightLoop.cs
@@ -819,10 +819,11 @@ namespace UnityEngine.Rendering.HighDefinition
         void InitShadowSystem(HDRenderPipelineAsset hdAsset, RenderPipelineResources defaultResources)
         {
             m_ShadowInitParameters = hdAsset.currentPlatformRenderPipelineSettings.hdShadowInitParams;
-            m_ShadowManager = HDShadowManager.instance;
+            m_ShadowManager = new HDShadowManager();
             m_ShadowManager.InitShadowManager(
                 defaultResources,
                 m_ShadowInitParameters,
+                m_RenderGraph,
                 defaultResources.shaders.shadowClearPS
             );
         }
@@ -831,7 +832,7 @@ namespace UnityEngine.Rendering.HighDefinition
         {
             if (m_ShadowManager != null)
             {
-                m_ShadowManager.Dispose();
+                m_ShadowManager.Cleanup(m_RenderGraph);
                 m_ShadowManager = null;
             }
         }
@@ -1858,8 +1859,9 @@ namespace UnityEngine.Rendering.HighDefinition
 
         internal bool GetEnvLightData(CommandBuffer cmd, HDCamera hdCamera, in ProcessedProbeData processedProbe, ref EnvLightData envLightData)
         {
-            // For the generic case, we consider that all probes have mip maps. The more specific case will override this attribute.
+            // By default, rough reflections are enabled for both types of probes.
             envLightData.roughReflections = 1.0f;
+            envLightData.distanceBasedRoughness = 0.0f;
 
             Camera camera = hdCamera.camera;
             HDProbe probe = processedProbe.hdProbe;
@@ -1962,6 +1964,9 @@ namespace UnityEngine.Rendering.HighDefinition
                     );
                     capturePosition = cameraPositionSettings.position;
                     envLightData.rangeCompressionFactorCompensation = Mathf.Max(probe.rangeCompressionFactor, 1e-6f);
+
+                    // Propagate the distance based information to the env light data (only if we are not an infinite projection)
+                    envLightData.distanceBasedRoughness = probe.settings.distanceBasedRoughness && !probe.isProjectionInfinite ? 1.0f : 0.0f;
 
                     break;
                 }